[Fix] Member captureDate 프로퍼티 추가에 따른 SwiftData MigrationPlan 추가
