+++ MC2-OneShot.xcodeproj/xcshareddata/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "43CDC966-40A7-4F4F-A7A9-985D8D2A8033"
+   type = "4"
+   version = "2.0">
+</Bucket>
